+++ CHANGELOG.md
@@ -26,6 +26,7 @@
 - Enable compression for forwarded uploads. ([#5965](https://github.com/getsentry/relay/pull/5965))
 - Change the default partitioning for the envelope buffer from semantic to round-robin. ([#5967](https://github.com/getsentry/relay/pull/5967))
 - Enable retries for upload requests to upstream. ([#5975](https://github.com/getsentry/relay/pull/5975))
+- Unconditionally create a trace context with a trace id for errors. ([#5979](https://github.com/getsentry/relay/pull/5979))
 
 ## 26.4.2